restructured tables in tables dir
