+++ src/tasks/hooks/useProjectTasksListData.ts
@@ -3,7 +3,7 @@ import { useCurrentUser } from "src/users/hooks/useCurrentUser"
 import getTasks, { GetTasksInput } from "../queries/getTasks"
 import { MemberPrivileges } from "@prisma/client"
 import { useMemberPrivileges } from "src/projectprivileges/components/MemberPrivilegesContext"
-import { processProjectTasks } from "../utils/processTasks"
+import { processProjectTasks } from "../tables/processing/processProjectTasks"
 
 export default function useProjecTasksListData(projectId: number | undefined) {
   const currentUser = useCurrentUser()
